requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-groq
langchain-huggingface
langchain-community
langgraph
chromadb
sentence-transformers
pypdf
python-docx
openpyxl
python-pptx
pandas
unstructured[all-docs]
python-magic-bin
openai-whisper
moviepy
opencv-python
pillow
torch
torchvision
torchaudio
beautifulsoup4
